+++ app/react-native/src/preview/components/StoryListView/index.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/destructuring-assignment */
 import React, { Component } from 'react';
 import { SectionList, TextInput, TouchableOpacity, View, SafeAreaView } from 'react-native';
 import styled from '@emotion/native';
@@ -158,7 +157,10 @@ export default class StoryListView extends Component<Props, State> {
   }
 
   render() {
-    const selectedStory = this.props.stories.getSelection();
+    const { stories } = this.props;
+    const { storyId } = stories.getSelection();
+    const selectedStory = stories.fromId(storyId);
+
     const { data } = this.state;
 
     return (